endeavour: separate collection of logs
When an application hangs, we need a utility that will collect client
and server logs after a script hangs and is killed. Thus, separate the
functionality to collect all logs into a separate script and replace the
current collection of logs code in run_client.sh and
run_client_resnet.sh with a call to this script.

Signed-off-by: James Nunez <james.nunez@intel.com>