endeavour: create collection of logs utility
DAOS client and server logs are collected after an application
completes. If that applciation hangs, is killed, or the script doesn't
complete, logs are not collected. In these cases, we need a utility to
collect logs.

Let's separate the functionality to collect all logs into a script and
replace the code to collect logs in run_client.sh and
run_client_resnet.sh with a call to this script.

Signed-off-by: James Nunez <james.nunez@intel.com>