endeavour: remove CLIENT/SERVER install dirs
Endeavour previously had different one OS on the clients and a different
OS on the servers. Endeavour recently changed to have the same OS on
clients and servers and, thus, we no longer need to have separate
CLIENT/install and SERVER/install build directories. We can remove the
different DAOS install directories in the scripts.

Signed-off-by: James Nunez <james.nunez@intel.com>