endeavour: add README.md in top edv directory
The Endeavour scripts have very little documentation on how to execute
them, what each script does and expected file/directory layout.  Add a
README.md to the top level directory of the Endeavour scripts as a
start to documenting the scripts and the expected organization of files.

More documentation is needed to allow someone new to run these
scripts on Endeavour.

Signed-off-by: James Nunez <james.nunez@intel.com>