frontera: get system name
There is a desire to run the test framework run on frontera on other
systems like Endeavour. A first step in this process is to figure out
what system we are running on. In the future, the system name will be
used to determine what steps are taken in the build process.

Signed-off-by: James A. Nunez <james.nunez@intel.com>