endeavour: allow use of pil4dfs IL
The Endeavour scripts currently only use the ioil Interception Library.
We need to allow for use of pil4dfs Interception Library.

The ‘-i’ command-line option will be redefined from the binary meaning
of use the interception library (1) or don’t use the library (0) to
specifying which interception library to use; ‘ioil’ or ‘pil4dfs’.

Signed-off-by: James A. Nunez <james.nunez@intel.com>